refactor(crypto)!: make (encode/decode)Base(58/64)Check private
